Split workflow tests into 2 files
